+++ MOVEit.MIT/Private/Globals.ps1
@@ -8,3 +8,8 @@ $script:Token = @()
 
 # User-Agent string to use
 $script:UserAgent = "MOVEit API"
+
+# Flag to SkipCertificateCheck.
+# Only supported in PowerShell v6 and later.
+# Will be set by Connect-MITServer
+$script:SkipCertificateCheck = $false