+++ MOVEit.MIT/Public/Invoke-MITWebRequest.ps1
@@ -34,6 +34,12 @@ function Invoke-MITWebRequest {
         $irmParams = @{}
         if ($Query) { $irmParams['Body'] = $Query}
 
+        # Add SkipCertificateCheck parameter if set
+        if ($script:SkipCertificateCheck) {
+            $irmParams['SkipCertificateCheck'] = $true
+            Write-Verbose "SkipCertificateCheck: $true"
+        }
+
         # If this is PowerShell 7, lets add the -SkipHttpErrorCheck param
         if ($PSVersionTable.PSVersion.Major -ge 7) { $irmParams['SkipHttpErrorCheck'] = $true}